[worker] refactor request creation into its own file
